split the main into 3 independent tasks
