+++ client/src/lib.rs
@@ -7,3 +7,7 @@ pub mod utils;
 pub use config::Config;
 pub use error::{ClientError, Result};
 pub use utils::*;
+
+pub mod deploy;
+pub mod retrive_funds;
+pub mod verify;
